requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
mysql==0.0.2
mysqlclient @ file:///C:/Users/jose_/Desktop/django/proyecto/entorno/users/Scripts/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
pytz==2020.4
sqlparse==0.4.1
Unipath==1.1
